Move Gravatar-related components under ./profile-gravatar
