+++ client/dashboard/me/profile-gravatar/index.tsx
@@ -16,8 +16,8 @@ import { createInterpolateElement, useMemo } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { useState } from 'react';
 import { SectionHeader } from '../../components/section-header';
-import EditGravatar from '../edit-gravatar';
-import GravatarLogo from '../profile/gravatar-logo';
+import EditGravatar from './edit-gravatar';
+import GravatarLogo from './gravatar-logo';
 import type { UserProfile } from '@automattic/api-core';
 import type { Field, Form } from '@wordpress/dataviews';
